feat: add WSS tracker support and fix async DNS resolution
- Add WebSocket Secure (wss://) tracker support for WebTorrent compatibility
- Fix UDP tracker DNS resolution to use async tokio::net::lookup_host()
- Add protocol module for boundary types
- Update documentation with WSS tracker protocol details
- Add CHANGELOG.md for version tracking
- Bump version to 0.1.4
