+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "gosh-dl"
-version = "0.1.3"
+version = "0.1.4"
 edition = "2021"
 rust-version = "1.75"
 description = "A fast, embeddable download engine for Rust. HTTP/HTTPS with multi-connection acceleration and full BitTorrent protocol support."
@@ -57,6 +57,12 @@ mainline = "6"
 
 # Multicast sockets (for LPD)
 socket2 = { version = "0.5", features = ["all"] }
+
+# WebSocket (for WSS trackers)
+tokio-tungstenite = { version = "0.24", features = ["rustls-tls-webpki-roots"] }
+
+# Base64 encoding
+base64 = "0.22"
 
 # Error handling
 thiserror = "2"
